Fix message payload and analytics
