+++ src/components/Highlights/TopIssues/highlightCards/HotSpotHighlightCard/index.tsx
@@ -1,12 +1,14 @@
 import { Row, createColumnHelper } from "@tanstack/react-table";
 import { useContext } from "react";
+import { sendUserActionTrackingEvent } from "../../../../../utils/actions/sendUserActionTrackingEvent";
 import { ConfigContext } from "../../../../common/App/ConfigContext";
 import { Table } from "../../../common/Table";
 import { TableText } from "../../../common/TableText";
+import { handleEnvironmentTableRowClick } from "../../../handleEnvironmentTableRowClick";
+import { trackingEvents } from "../../../tracking";
 import { HighlightCard } from "../../common/HighlightCard";
 import { EnvironmentData, HotSpotMetrics } from "../../types";
 import { addEnvironmentColumns } from "../addEnvironmentColumns";
-import { handleEnvironmentTableRowClick } from "../goToEnvironmentIssues";
 import { HotSpotHighlightCardProps } from "./types";
 
 export const HotSpotHighlightCard = ({ data }: HotSpotHighlightCardProps) => {
@@ -28,10 +30,15 @@ export const HotSpotHighlightCard = ({ data }: HotSpotHighlightCardProps) => {
   const columns = addEnvironmentColumns(columnHelper, metricsColumns);
 
   const handleTableRowClick = (row: Row<EnvironmentData<HotSpotMetrics>>) => {
+    sendUserActionTrackingEvent(
+      trackingEvents.TOP_ISSUES_CARD_TABLE_ROW_CLICKED,
+      {
+        insightType: data.insightType
+      }
+    );
     handleEnvironmentTableRowClick(
       config.environments,
-      row.original.environmentName,
-      data.insightType
+      row.original.environmentName
     );
   };
