+++ src/components/Highlights/TopIssues/highlightCards/SpanEndpointBottleneckHighlightCard/index.tsx
@@ -1,15 +1,17 @@
 import { Row, createColumnHelper } from "@tanstack/react-table";
 import { useContext } from "react";
 import { Duration } from "../../../../../globals";
+import { sendUserActionTrackingEvent } from "../../../../../utils/actions/sendUserActionTrackingEvent";
 import { getDurationString } from "../../../../../utils/getDurationString";
 import { ConfigContext } from "../../../../common/App/ConfigContext";
 import { Table } from "../../../common/Table";
 import { TableTag } from "../../../common/TableTag";
 import { TableText } from "../../../common/TableText";
+import { handleEnvironmentTableRowClick } from "../../../handleEnvironmentTableRowClick";
+import { trackingEvents } from "../../../tracking";
 import { HighlightCard } from "../../common/HighlightCard";
 import { EnvironmentData, SpanEndpointBottleneckMetrics } from "../../types";
 import { addEnvironmentColumns } from "../addEnvironmentColumns";
-import { handleEnvironmentTableRowClick } from "../goToEnvironmentIssues";
 import { SpanEndpointBottleneckHighlightCardProps } from "./types";
 
 export const SpanEndpointBottleneckHighlightCard = ({
@@ -63,10 +65,15 @@ export const SpanEndpointBottleneckHighlightCard = ({
   const handleTableRowClick = (
     row: Row<EnvironmentData<SpanEndpointBottleneckMetrics>>
   ) => {
+    sendUserActionTrackingEvent(
+      trackingEvents.TOP_ISSUES_CARD_TABLE_ROW_CLICKED,
+      {
+        insightType: data.insightType
+      }
+    );
     handleEnvironmentTableRowClick(
       config.environments,
-      row.original.environmentName,
-      data.insightType
+      row.original.environmentName
     );
   };
